Add inventory_bills stream
