+++ tap_service_titan/streams/accounting.py
@@ -432,3 +432,206 @@ class PaymentsStream(ServiceTitanExportStream):
     def path(self) -> str:
         """Return the API path for the stream."""
         return f"/accounting/v2/tenant/{self._tap.config['tenant_id']}/export/payments"
+
+
+class InventoryBillsStream(ServiceTitanExportStream):
+    """Define payments stream."""
+
+    name = "inventory_bills"
+    primary_keys: t.ClassVar[list[str]] = ["id"]
+    replication_key: str = "modifiedOn"
+    schema = th.PropertiesList(
+        th.Property("id", th.IntegerType),
+        th.Property("syncStatus", th.StringType),
+        th.Property("referenceNumber", th.StringType),
+        th.Property("date", th.DateTimeType),
+        th.Property("type", th.StringType),
+        th.Property("typeId", th.StringType),
+        th.Property("total", th.StringType),
+        th.Property("unappliedAmount", th.StringType),
+        th.Property("memo", th.StringType),
+        th.Property(
+            "customer",
+            th.ObjectType(
+                th.Property("id", th.IntegerType), th.Property("name", th.StringType)
+            ),
+        ),
+        th.Property(
+            "businessUnit",
+            th.ObjectType(
+                th.Property("id", th.IntegerType), th.Property("name", th.StringType)
+            ),
+        ),
+        th.Property(
+            "batch",
+            th.ObjectType(
+                th.Property("id", th.IntegerType),
+                th.Property("number", th.StringType),
+                th.Property("name", th.StringType),
+            ),
+        ),
+        th.Property("createdBy", th.StringType),
+        th.Property(
+            "generalLedgerAccount",
+            th.ObjectType(
+                th.Property("id", th.IntegerType),
+                th.Property("name", th.StringType),
+                th.Property("number", th.StringType),
+                th.Property("type", th.StringType),
+                th.Property("detailType", th.StringType),
+            ),
+        ),
+        th.Property(
+            "appliedTo",
+            th.ArrayType(
+                th.ObjectType(
+                    th.Property("appliedId", th.IntegerType),
+                    th.Property("appliedTo", th.IntegerType),
+                    th.Property("appliedAmount", th.StringType),
+                    th.Property("appliedOn", th.DateTimeType),
+                    th.Property("appliedBy", th.StringType),
+                    th.Property("appliedToReferenceNumber", th.StringType),
+                )
+            ),
+        ),
+        th.Property(
+            "customFields",
+            th.ArrayType(
+                th.ObjectType(
+                    th.Property("name", th.StringType),
+                    th.Property("value", th.StringType),
+                )
+            ),
+        ),
+        th.Property("authCode", th.StringType),
+        th.Property("checkNumber", th.StringType),
+        th.Property("modifiedOn", th.DateTimeType),
+        th.Property("createdOn", th.DateTimeType),
+        th.Property("active", th.BooleanType),
+    ).to_dict()
+
+    @cached_property
+    def path(self) -> str:
+        """Return the API path for the stream."""
+        return f"/accounting/v2/tenant/{self._tap.config['tenant_id']}/export/payments"
+
+
+class InventoryBillsStream(ServiceTitanExportStream):
+    """Define inventory bills stream."""
+
+    name = "inventory_bills"
+    primary_keys: t.ClassVar[list[str]] = ["id"]
+    replication_key: str = "createdOn"
+    schema = th.PropertiesList(
+        th.Property("id", th.IntegerType),
+        th.Property("purchaseOrderId", th.IntegerType),
+        th.Property("syncStatus", th.StringType),
+        th.Property("referenceNumber", th.StringType),
+        th.Property("vendorNumber", th.StringType),
+        th.Property("summary", th.StringType),
+        th.Property("billDate", th.DateTimeType),
+        th.Property("billAmount", th.StringType),
+        th.Property("taxAmount", th.StringType),
+        th.Property("shippingAmount", th.StringType),
+        th.Property("createdBy", th.StringType),
+        th.Property("createdOn", th.DateTimeType),
+        th.Property("termName", th.StringType),
+        th.Property("dueDate", th.DateTimeType),
+        th.Property("shipToDescription", th.StringType),
+        th.Property(
+            "batch",
+            th.ObjectType(
+                th.Property("id", th.IntegerType),
+                th.Property("number", th.StringType),
+                th.Property("name", th.StringType),
+            ),
+        ),
+        th.Property(
+            "businessUnit",
+            th.ObjectType(
+                th.Property("id", th.IntegerType),
+                th.Property("name", th.StringType),
+            ),
+        ),
+        th.Property(
+            "vendor",
+            th.ObjectType(
+                th.Property("id", th.IntegerType),
+                th.Property("name", th.StringType),
+            ),
+        ),
+        th.Property(
+            "shipTo",
+            th.ObjectType(
+                th.Property("street", th.StringType),
+                th.Property("unit", th.StringType),
+                th.Property("city", th.StringType),
+                th.Property("state", th.StringType),
+                th.Property("zip", th.StringType),
+                th.Property("country", th.StringType),
+            ),
+        ),
+        th.Property(
+            "items",
+            th.ArrayType(
+                th.ObjectType(
+                    th.Property("order", th.IntegerType),
+                    th.Property("name", th.StringType),
+                    th.Property("description", th.StringType),
+                    th.Property("quantity", th.StringType),
+                    th.Property("cost", th.StringType),
+                    th.Property("inventoryLocation", th.StringType),
+                    th.Property("serialNumber", th.StringType),
+                    th.Property(
+                        "generalLedgerAccount",
+                        th.ObjectType(
+                            th.Property("id", th.IntegerType),
+                            th.Property("name", th.StringType),
+                            th.Property("number", th.StringType),
+                            th.Property("type", th.StringType),
+                            th.Property("detailType", th.StringType),
+                        ),
+                    ),
+                    th.Property(
+                        "costOfSaleAccount",
+                        th.ObjectType(
+                            th.Property("id", th.IntegerType),
+                            th.Property("name", th.StringType),
+                            th.Property("number", th.StringType),
+                            th.Property("type", th.StringType),
+                            th.Property("detailType", th.StringType),
+                        ),
+                    ),
+                    th.Property(
+                        "assetAccount",
+                        th.ObjectType(
+                            th.Property("id", th.IntegerType),
+                            th.Property("name", th.StringType),
+                            th.Property("number", th.StringType),
+                            th.Property("type", th.StringType),
+                            th.Property("detailType", th.StringType),
+                        ),
+                    ),
+                    th.Property("skuId", th.IntegerType),
+                    th.Property("skuCode", th.StringType),
+                )
+            ),
+        ),
+        th.Property(
+            "customFields",
+            th.ArrayType(
+                th.ObjectType(
+                    th.Property("name", th.StringType),
+                    th.Property("value", th.StringType),
+                )
+            ),
+        ),
+        th.Property("jobId", th.IntegerType),
+        th.Property("jobNumber", th.StringType),
+        th.Property("active", th.BooleanType),
+    ).to_dict()
+
+    @cached_property
+    def path(self) -> str:
+        """Return the API path for the stream."""
+        return f"/accounting/v2/tenant/{self._tap.config['tenant_id']}/export/inventory-bills"
